pyproject.toml
[tool.poetry]
name = "nevod-pulse"
version = "0.1.0"
description = "nevod api and monitoring system"
authors = ["Pavel <Plyubimov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
prometheus-client = "^0.21.0"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.1"
sqlmodel = "^0.0.22"
backoff = "^2.2.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
